backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
python-dotenv
google-generativeai
langchain
langchain-google-genai
supabase
PyMuPDF
pdfplumber
httpx
